fix rebuild bug

+++ src/index.ts
@@ -24,26 +24,30 @@ export default (
       mockDir: joi.string(),
     });
   });
-
-  ctx.onBuildFinish(async () => {
-    const { appPath } = ctx.paths;
-    const port = pluginOpts.port || 9527;
-    const host = pluginOpts.host || '0.0.0.0';
-    const basePath = pluginOpts.basePath || '/';
-    const mockDir = pluginOpts.mockDir || 'mock';
-    await kill(port, 'tcp');
-    const app = express();
-    app.use(bodyParser.json());
-    app.use(mockjsMiddleware(path.join(appPath, mockDir)));
-    app.listen(port, host, () => {
-      console.log(
-        chalk.green(
-          `数据 mock 服务已启动，Server 地址 http://${host}:${port}${basePath}, mock文件夹为 ${path.join(
-            appPath,
-            mockDir
-          )}`
-        )
-      );
-    });
+  let isFirstWatch = true;
+  ctx.onBuildFinish(async ({ isWatch }) => {
+    let needStart = !isWatch || isFirstWatch;
+    if (needStart) {
+      const { appPath } = ctx.paths;
+      const port = pluginOpts.port || 9527;
+      const host = pluginOpts.host || '0.0.0.0';
+      const basePath = pluginOpts.basePath || '/';
+      const mockDir = pluginOpts.mockDir || 'mock';
+      await kill(port, 'tcp');
+      const app = express();
+      app.use(bodyParser.json());
+      app.use(mockjsMiddleware(path.join(appPath, mockDir)));
+      app.listen(port, host, () => {
+        console.log(
+          chalk.green(
+            `数据 mock 服务已启动，Server 地址 http://${host}:${port}${basePath}, mock文件夹为 ${path.join(
+              appPath,
+              mockDir
+            )}`
+          )
+        );
+      });
+    }
+    isFirstWatch = false;
   });
 };
